rename app.py to error_handler
