+++ ckanext/feedback/views/download.py
@@ -1,7 +1,7 @@
 from flask import Blueprint
 
 from ckanext.feedback.controllers.download import DownloadController
-from ckanext.feedback.views.app import add_error_handler
+from ckanext.feedback.views.error_handler import add_error_handler
 
 blueprint = Blueprint(
     'download',
